+++ middleware.go
@@ -8,6 +8,7 @@ import (
 	"log"
 	"net/http"
 	"runtime/debug"
+	"strings"
 	"sync"
 )
 
@@ -140,3 +141,36 @@ func RecoveryLogger(debug bool, logger *log.Logger) MiddlewareFunc {
 
 // Skipper is a function that indicates whether current request is skippable.
 type Skipper func(ctx *Context) bool
+
+// PathSkipper returns skipper with the given patterns.
+// Pattern has two forms, one is that contains a certain path, another contains a wildcard,
+// both of them are case-insensitive.
+//   Pattern     Path            Skippable
+//   ""          "/"             false
+//   "/"         "/"             true
+//   "/"         "/login"        false
+//   "/login"    "/login"        true
+//   "/login"    "/Login"        true
+//   "/login"    "/LOGIN"        true
+//   "/guest*"   "/guest"        true
+//   "/guest*"   "/guest/foo"    true
+//   "/guest*"   "/guest/bar"    true
+func PathSkipper(patterns ...string) Skipper {
+	return func(ctx *Context) bool {
+		for _, pattern := range patterns {
+			if pattern == "" {
+				continue
+			}
+			if pattern[len(pattern)-1] == '*' && len(ctx.Request.URL.Path) >= len(pattern)-1 {
+				length := len(pattern) - 1
+				if strings.EqualFold(ctx.Request.URL.Path[:length], pattern[:length]) {
+					return true
+				}
+			}
+			if strings.EqualFold(pattern, ctx.Request.URL.Path) {
+				return true
+			}
+		}
+		return false
+	}
+}