+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 
 under development
 -----------------
+- Add `Skipper`.
 
 v1.11.0 April 15, 2020
 ----------------------